pyproject.toml
[tool.poetry]
name = "smsa-server"
version = "0.1.0"
description = ""
authors = ["Daniel Tse <daniel.tse1211@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
numpy = "^2.2.1"
pandas = "^2.2.3"
scipy = "^1.14.1"
scikit-learn = "^1.6.0"
spacy = "^3.8.3"
joblib = "^1.4.2"
asyncpraw = "^7.8.1"
gcloud-aio-storage = "^9.3.0"
gcloud-rest-storage = "^9.3.0"
fastapi = {extras = ["standard"], version = "^0.115.6"}
pydantic-settings = "^2.7.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
flake8 = "^7.1.1"
isort = "^5.13.2"
black = {extras = ["jupyter"], version = "^24.10.0"}
ipython = "^8.31.0"
matplotlib = "^3.10.0"
seaborn = "^0.13.2"
mlflow = "^2.19.0"
pre-commit = "^4.0.1"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.0"
pytest-dotenv = "^0.5.2"
flake8-pyproject = "^1.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
skip-string-normalization = true

[tool.isort]
multi_line_output = 3
length_sort = true
skip_gitignore = true
include_trailing_comma = true

[tool.flake8]
extend-exclude = [".git", ".venv", "__pycache__", "*.ipynb"]
max-line-length = 88
count = true

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_default_fixture_loop_scope = "function"
